Most thing re-rewritten
The plotting code was re-written. I believe now it will be easier to plot new data.
I am using a 3x2 plot. That is, one window with 6 plotting areas. This way we don't
need to create something new for every new data.

There is also a parse input for the flow computer. The user can call:
python main.py flowComputer=rangePercent_regular
other options to flowComputer are: regular, rangePercent_windowFlow, windowFlow

bug to be fixed: windowFlow returns empty array whenever it is not ready to output something.
I have to deal with it on the main program. The program runs fine for the regular computer.